+++ .gitignore
@@ -1,7 +1,15 @@
+Thumbs.db
 .DS_Store
+.gradle
+build/
+out/
+.idea
 *.iml
 *.ipr
 *.iws
+.project
+.settings
+.classpath
 *.log
 /grails-crypto-*.zip
 /grails-crypto-*.zip.sha1
