--- application.properties
@@ -1,2 +1,0 @@
-app.grails.version=2.5.5
-app.name=crypto
