+++ gradle.properties
@@ -1,0 +1,5 @@
+grailsVersion=4.0.11
+gorm.version=7.0.8.RELEASE
+org.gradle.daemon=true
+org.gradle.parallel=true
+org.gradle.jvmargs=-Dfile.encoding=UTF-8 -Xmx1024M
