+++ grails-app/init/cr/co/arquetipos/Application.groovy
@@ -1,0 +1,13 @@
+package cr.co.arquetipos
+
+
+import grails.boot.GrailsApp
+import grails.boot.config.GrailsAutoConfiguration
+import grails.plugins.metadata.PluginSource
+
+@PluginSource
+class Application extends GrailsAutoConfiguration {
+    static void main(String[] args) {
+        GrailsApp.run(Application, args)
+    }
+}
