+++ grails-app/init/cr/co/arquetipos/BootStrap.groovy
@@ -1,0 +1,9 @@
+package cr.co.arquetipos
+
+class BootStrap {
+
+    def init = { servletContext ->
+    }
+    def destroy = {
+    }
+}
