--- wrapper/grails-wrapper.properties
@@ -1,1 +1,0 @@
-wrapper.dist.url=https://github.com/grails/grails-core/releases/download/v2.5.5/
